refactor(any-screen): 길찾기 및 공유 관련 컴포넌트 분리 및 정,동적 화면에 적용
- 길찾기 및 공유 기능을 묶은 LocationActionButtons 컴포넌트를 새로 추가하고, Google 지도 경로 열기와 Share API 호출을 처리하도록 구현

- 정적 결과 화면에서 기존 버튼 코드를 제거하고 새 컴포넌트를 활용해 버튼을 표시하도록 리팩터링

- 동적 결과 화면의 각 장소 항목에 동일한 컴포넌트를 배치해 길찾기 및 공유 기능을 지원하고, 레이아웃 간격을 조정

- 홈 화면에서 하드코딩된 핀 배열을 삭제해 정적 마커가 표시되지 않도록 수정
